pyproject.toml
[build-system]
requires = [
    "setuptools",
    "robotframework",
    ]
build-backend = "setuptools.build_meta"

[project]
name = "robotframework-TrainingREST"
authors = [{name="Artur Ziolkowksi", email="artur.k.ziolkowski@gmail.com"},
]
description = "Go REST Library for Robot Framework"
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "robotframework",
    "setuptools",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = {text = "Apache License 2.0"}
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/ajadach/robotframework-TrainingREST"
"Keyword docs" = "https://github.com/ajadach/robotframework-TrainingREST"
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120